memory: AddressConverter type to adapt buses with different address types
This is important to allow using a Memory (whose address type is always
usize) in a CPU whose address type is either u32, u64, or u128.

However, this isn't quite right yet because it assumes conversion to usize
is possible for all those types. In practice, there isn't a guaranteed
conversion from u64 to usize because that would truncate on 32-bit
targets, so in a subsequent change we'll switch this over to using TryInto
and handling the conversion error dynamically.