hart: Trait for implementations of hart state
This is intended to encapsulate the state for a hart and expose operations
that allow callers to mutate that state while maintaining invariants
expected by whatever ISAs and ISA extensions the hart implements.

A separate component will wrap this with an implementation of all of
the relevant RISC-V instructions in terms of these operations, which
can then in turn be wrapped by an execution driver that could be as simple
as just executing one instruction at a time in a loop, or could be a
more complex debug interface that allows more fine-grain manipulation
of the hart.