memory: AddressTransformer bus implementation
This is a simple Bus wrapper that pre-checks and transforms incoming
addresses before passing them on to the wrapped Bus.

We'll probably have a more sophisticated mapping mechanism later that can
also select between multiple different underlying buses, but this is a
simple place to start to handle the situation where we're loading a memory
image from disk that expects to appear at a particular address, so we can
offset accesses to that address into the range expected by the memory
object.